pyproject.toml
[project]
name = "agentic-phidata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckduckgo-search>=7.2.1",
    "fastapi>=0.115.6",
    "groq>=0.13.1",
    "packaging>=24.2",
    "phidata>=2.7.6",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.34.0",
    "yfinance>=0.2.51",
]
